backend/requirements.txt
Django==5.0.6
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
django-cors-headers==4.4.0
mysqlclient==2.2.4
celery==5.4.0
redis==5.0.7
python-dotenv==1.0.1
requests==2.31.0
reportlab==4.0.4
openpyxl==3.1.2
Pillow==10.0.0
pyotp==2.9.0
qrcode[pil]==7.4.2
sslyze==5.2.0
python-libnmap==0.7.3
django-ratelimit==4.1.0
